09-Project/requirements.txt
requests
pyfiglet
pytest
python-dotenv
tabulate